rename pgxplugin -> pgx
