+++ plugin/postgres/pgx/templates.go
@@ -1,4 +1,4 @@
-package pgxplugin
+package pgx
 
 // init
